refactor: add routes directory
