+++ src/routes/albums/albums.module.ts
@@ -1,8 +1,8 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { AlbumsService } from './albums.service';
 import { AlbumsController } from './albums.controller';
-import { TracksModule } from 'src/tracks/tracks.module';
-import { FavotitesModule } from 'src/favotites/favotites.module';
+import { TracksModule } from 'src/routes/tracks/tracks.module';
+import { FavotitesModule } from 'src/routes/favotites/favotites.module';
 import { AlbumsRepository } from 'src/db/albums.repository';
 
 @Module({
