+++ src/routes/albums/entities/album.entity.ts
