+++ src/routes/artists/artists.controller.ts
