+++ src/routes/artists/artists.module.ts
@@ -1,9 +1,9 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { ArtistsService } from './artists.service';
 import { ArtistsController } from './artists.controller';
-import { AlbumsModule } from 'src/albums/albums.module';
-import { TracksModule } from 'src/tracks/tracks.module';
-import { FavotitesModule } from 'src/favotites/favotites.module';
+import { AlbumsModule } from 'src/routes/albums/albums.module';
+import { TracksModule } from 'src/routes/tracks/tracks.module';
+import { FavotitesModule } from 'src/routes/favotites/favotites.module';
 import { ArtistsRepository } from 'src/db/artists.repository';
 
 @Module({
