+++ src/routes/artists/dto/create-artist.dto.ts
