+++ src/routes/artists/dto/update-artist.dto.ts
