+++ src/routes/artists/entities/artist.entity.ts
