+++ src/routes/artists/model/artist.model.ts
