+++ src/routes/favotites/entities/favotite.entity.ts
