+++ src/routes/favotites/favotites.module.ts
@@ -1,9 +1,9 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { FavotitesService } from './favotites.service';
 import { FavotitesController } from './favotites.controller';
-import { ArtistsModule } from 'src/artists/artists.module';
-import { AlbumsModule } from 'src/albums/albums.module';
-import { TracksModule } from 'src/tracks/tracks.module';
+import { ArtistsModule } from 'src/routes/artists/artists.module';
+import { AlbumsModule } from 'src/routes/albums/albums.module';
+import { TracksModule } from 'src/routes/tracks/tracks.module';
 import { FavoritesRepository } from 'src/db/favorites.repository';
 
 @Module({
