+++ src/routes/tracks/tracks.module.ts
@@ -1,7 +1,7 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { TracksService } from './tracks.service';
 import { TracksController } from './tracks.controller';
-import { FavotitesModule } from 'src/favotites/favotites.module';
+import { FavotitesModule } from 'src/routes/favotites/favotites.module';
 import { TracksRepository } from 'src/db/tracks.repository';
 
 @Module({
