+++ src/routes/user/model/user.model.ts
